Remove display contents tags
