+++ packages/sdk-components-react-radix/src/tooltip.tsx
@@ -2,10 +2,6 @@
 // We can't use .displayName until this is merged https://github.com/styleguidist/react-docgen-typescript/pull/449
 
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
-import {
-  splitPropsWithWebstudioAttributes,
-  type WebstudioAttributes,
-} from "@webstudio-is/react-sdk";
 
 import {
   forwardRef,
@@ -23,19 +19,9 @@ type BuilderTooltipProps = {
   isOpen: "initial" | "open" | "closed";
 };
 
-/**
- * Tooltip and TooltipTrigger are HTML-less components.
- * To make them work in our system, we wrap their attributes with a div that has a display: contents property.
- *
- * These divs function like fragments, with all web studio-related attributes attached to them.
- */
-const DisplayContentsStyle = { display: "contents" };
-
 export const Tooltip = forwardRef<
   ElementRef<"div">,
-  WebstudioAttributes &
-    ComponentPropsWithoutRef<typeof TooltipPrimitive.Root> &
-    BuilderTooltipProps
+  ComponentPropsWithoutRef<typeof TooltipPrimitive.Root> & BuilderTooltipProps
 >(({ open: openProp, isOpen, ...props }, ref) => {
   const open =
     openProp ??
@@ -55,19 +41,15 @@ export const Tooltip = forwardRef<
  * which would prevent us from displaying styles properly in the builder.
  */
 export const TooltipTrigger = forwardRef<
-  ElementRef<"div">,
-  WebstudioAttributes & { children: ReactNode }
+  ElementRef<"button">,
+  { children: ReactNode }
 >(({ children, ...props }, ref) => {
   const firstChild = Children.toArray(children)[0];
-  const [webstudioAttributes, restProps] =
-    splitPropsWithWebstudioAttributes(props);
 
   return (
-    <div ref={ref} style={DisplayContentsStyle} {...webstudioAttributes}>
-      <TooltipPrimitive.Trigger asChild={true} {...restProps}>
-        {firstChild ?? <button>Add button or link</button>}
-      </TooltipPrimitive.Trigger>
-    </div>
+    <TooltipPrimitive.Trigger asChild={true} ref={ref} {...props}>
+      {firstChild ?? <button>Add button or link</button>}
+    </TooltipPrimitive.Trigger>
   );
 });
